+++ src/iir/biquad.rs
@@ -49,7 +49,7 @@ use crate::Coefficient;
 ///
 /// The filter applies clamping such that `min <= y <= max`.
 ///
-/// See [`crate::iir::Filter`] and [`crate::iir::Pid`] for ways to generate coefficients.
+/// See [`crate::iir::Filter`] and [`crate::iir::PidBuilder`] for ways to generate coefficients.
 ///
 /// # Fixed point
 ///
@@ -100,7 +100,7 @@ use crate::Coefficient;
 ///   coefficients/offset sets.
 /// * Cascading multiple IIR filters allows stable and robust
 ///   implementation of transfer functions beyond biquadratic terms.
-#[derive(Copy, Clone, Debug, Deserialize, Serialize, PartialEq, PartialOrd)]
+#[derive(Clone, Debug, Deserialize, Serialize, PartialEq, PartialOrd)]
 pub struct Biquad<T> {
     ba: [T; 5],
     u: T,
@@ -134,7 +134,7 @@ where
     C: Float + AsPrimitive<T>,
 {
     fn from(ba: &[C; 6]) -> Self {
-        let ia0 = C::one() / ba[3];
+        let ia0 = ba[3].recip();
         Self::from([
             T::quantize(ba[0] * ia0),
             T::quantize(ba[1] * ia0),