+++ src/iir/mod.rs
@@ -6,3 +6,5 @@ mod coefficients;
 pub use coefficients::*;
 mod pid;
 pub use pid::*;
+mod repr;
+pub use repr::*;