+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [UNRELEASED](https://github.com/quartiq/idsp/compare/v0.17.0..HEAD) - DATE
+
+### Changed
+
+* `Filter`: use non-inverse Q
+* `Pid` renamed to `PidBuilder`
+* `Pid` added with miniconf support and just P,I,D gains, limits, setpoint, min, max
+* `BiquadRepr` enum to convert from different representations to raw biquad
+
 ## [0.17.0](https://github.com/quartiq/idsp/compare/v0.16.0..v0.17.0) - 2025-01-20
 
 ### Changed